消息队列销峰秒杀流量

+++ project-cluster-service/cloud-rear-server/seckillservice/src/main/java/org/jnyou/seckillservice/vo/SeckillSkuVo.java
@@ -41,7 +41,7 @@ public class SeckillSkuVo {
     /**
      * 每人限购数量
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * 排序
      */
